src/app/papi/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
Authlib==1.3.0
bcrypt==3.2.0
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
command-not-found==0.3
contourpy==1.2.0
cryptography==42.0.3
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.9.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
duplicity==0.8.21
EasyProcess==1.1
entrypoint2==1.1
evdev==1.6.1
exceptiongroup==1.2.0
fasteners==0.14.1
ffmpeg==1.4
filelock==3.13.1
Flask==2.2.2
fonttools==4.44.3
future==0.18.2
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-speech==2.24.1
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.3
httplib2==0.20.2
httpx==0.26.0
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
matplotlib==3.8.1
meson==0.61.2
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mss==9.0.1
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
openai==1.12.0
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.0.1
platformdirs==4.1.0
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.11
PyAutoGUI==0.9.54
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.6.1
pydantic_core==2.16.2
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.6.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyscreenshot==3.1
PyScreeze==0.1.29
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
rsa==4.9
scipy==1.12.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
systemd-python==234
tqdm==4.66.2
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.0
usb-creator==0.3.7
virtualenv==20.25.0
wadllib==1.3.6
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
zipp==1.0.0
